+++ vector/src/main/java/im/vector/app/core/di/VectorModule.kt
@@ -34,6 +34,7 @@ import im.vector.app.features.ui.UiStateRepository
 import org.matrix.android.sdk.api.Matrix
 import org.matrix.android.sdk.api.auth.AuthenticationService
 import org.matrix.android.sdk.api.legacy.LegacySessionImporter
+import org.matrix.android.sdk.api.raw.RawService
 import org.matrix.android.sdk.api.session.Session
 
 @Module
@@ -77,6 +78,12 @@ abstract class VectorModule {
         @JvmStatic
         fun providesAuthenticationService(matrix: Matrix): AuthenticationService {
             return matrix.authenticationService()
+        }
+
+        @Provides
+        @JvmStatic
+        fun providesRawService(matrix: Matrix): RawService {
+            return matrix.rawService()
         }
     }